Speed the Lua connection handling CRC code by about 30% by removing some unnecessary logical operations and using a polynomial lookup table instead of computing the CRC polynomial transformation for every byte.
The binary operations available in Lua are horrendously slow.  Most likely due to the conversion from double to integer on entry and integer to double on exit.

A quick toss together in C seemed to indicate a thousand fold performance improvement is possible in this integer and bit operation heavy CRC function.