Further fine tuning of the Lua CRC code.
By noticing that some logical operations correspond nicely to floating point operations, it is possible to save time.  The key is the avoidance of the conversion to integer and conversion back.

Specifically:

    a left shift is a multiply;
    a right shift is the floor of a reciprocal multiply and
    the bitwise AND operations that are used correspond to a floating point modulus.

The latter two of these involve expensive floating point operations (floor and modulus) but they seem to be cheaper than the conversions to integer (which are essentially the same but done twice).

Right shift is possibly better thought of as a floor of a divide, however division is very slow on many processors (especially ARM) so a reciprocal multiply works better.