Update timers so they take arguments in seconds not milliseconds as per Bug 40: http://big-gay-al.rinstrumau.local/show_bug.cgi?id=40
Remove the event queue from rinSystem and implement this functionality using the timer queue instead.  This guarantees that these events are executed when the system is otherwise idle which will prevent a rather subtle double processing of the file descriptors in the event of a recursive call to rinSystem's processEvents (which happens surprisingly often unfortunately).