Remove no longer used function from rinSystem.
Automatically detect the presence of posix support and if it is present, use the posix clock_gettime call in preference to the socket gettime.

The posix call is guaranteed monotonic regardless of system clock changes and it is usually far finer grained.  The socket call looks suspiciously like the time(2) system call which is coarse and suffers from system time and date changes in nasty ways.