Bug: http://big-gay-al.rinstrumau.local/show_bug.cgi?id=46
Implement mismatching table loads.
Add a return result code to the CSV loader that indicates what was done to load the file.

The common column detecting code isn't optimal.  Its time complexity is larger than need be.  If the CSV file is an exact match for what is wanted, this code isn't required -- this will likely be a very common case.