Move the buzzer control out of the keyboard handling file.
Add a basic unit test for the buzzer control.  This is segregated into two pieces.  The first validates the values of the constants so that if they ever change we'll know about it.  The second checks the low level messages being sent by the calls.

Begin an embroynic set register message verification module.