Change key names to non-magic codes.
Digits are represented by the numbers 0 - 9.
The rest of the keys are represented by strings.

Updated luadoc to match.

Updated examples to match.

Added a reverse register mapping in a separate table since that is far more sensible and like the way the mapping/unmapping has been done everywhere else.