Allow a more general form of user id.
The user can be specified by number, either a numeric value or a single character string.  A user can also be specified by register name (appropriate to each class of functions of course).

In the case of a numeric value, the floor is taken first and it is range checked.