Self checksumming of all the lua code.
Early in the start up process, every lua file included in the rinApp framework is read and the combined lump is checksummed using MD5.

The list of files is generated as part of the build process, although this file cannot be included in the checksumming process, changing it will almost certainly change the checksum itself.

The checksum for the entire suite is included in the package directory so comparisons are easy.