Extend the time parameter to the varous write to display calls so that in addition to a numeric time value, it can also take a table argument with fields or a string argument that is parsed into tabular form.
The fields that are available are:

time is the old time argument -- the number of seconds between fragments of a message.  It is also used at the end of the message for the wait and clear options.

once displays the message once.

wait displays the message once and pauses execution until it has been completely displayed.

clear displays the message once and clears the display afterwards.

Natually these options can be mixed and matched in any combination.