Deprecate saveBot/restoreBot in favour of saveBottom which returns a function that does the restore.  This allows these calls to properly nest which is a necessity for the menuing system.
Add lots of LDoc for the menuing.  Fix a few bugs and add in a few more features.

Add a menu example that uses each of the different items.

Factor out a null function to reduce resouce use a little.

Use the field table values for field functions instead of cached local copies.