Allow a USB keyboard or key pad to drive Lua scripts as if the front panel keys are being pressed.
The rinUSBKBD module is gone.  The library deals with the incoming key strokes and dispatches them via the display to normal handling.

Fix a minor problem with the key auto repeat.

Bump the version number.

At this point, the dialog functions have not been modified to work with USB keyboards.  This will be more problematic, especially editReg.