Add date and calendar support.
We can now work out the number of days between dates and what date is a specified number of days before or after another.

Additionally, month names and week day names can also be determined from a year, month and day trio.  These names and short forms are in English.